fix: debounce function
